L5 - AI Microservice (GCP Deployment)/requirements.txt
# Requirements for Google Cloud Run
Flask==2.0.1 # used for data types
gunicorn==20.1.0 # used to "be" a service

# Requirements for the service body
fastbook==0.0.18 # the only way we could figure out how to get fastai and most of its dependencies
google-cloud-storage==1.42.1 # use this package to fetch the model from Google Cloud Storage

spacy==3.3.1